requirements.txt
Flask
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
packaging
protobuf
grpcio-status
streamlit
onnxruntime
